+++ CHANGELOG.md
@@ -10,6 +10,7 @@ v0.15.0 (?? ??? 2018)
 - Parse formatting inside HTML lists - @troosan @samimussbach #1239 #945 #1215 #508
 - Parsing of CSS `direction` instruction, HTML `lang` attribute, formatting inside table cell - @troosan #1273 #1252 #1254
 - Add support for Track changes @Cip @troosan #354 #1262
+- Add support for fixed Table Layout @aoloe @ekopach @troosan #841 #1276
 
 ### Fixed
 - Fix reading of docx default style - @troosan #1238